% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/cleanup.R
\name{cleanup}
\alias{cleanup}
\title{cleanup.R}
\usage{
cleanup(x)
}
\arguments{
\item{x}{character}
}
\description{
cleanup.R
}
\examples{
cleanup(as.character(citation()) cleans and substitute citation()
cleanup(string) cleans and substitute string
}

