% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/filter_tokens.R
\name{filter_tokens}
\alias{filter_tokens}
\title{filter_tokens}
\usage{
filter_tokens(x, filters)
}
\arguments{
\item{x}{character}

\item{filters}{a char vector of tokens to exclude}
}
\description{
filter_tokens
}
\examples{
filter_words(as.character(citation()),filters) removes tokens contained in filters
filter_words(string,badwords) removes from string tokens contained in badwords
}

