% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/filtf.R
\name{filtf}
\alias{filtf}
\title{filtf.R}
\usage{
filtf(x, f)
}
\arguments{
\item{x}{character vector}

\item{f}{a char vector of tokens to exclude}
}
\description{
filtf.R
}
\examples{
filtf(string,badwords) removes from string vector regex matching items contained in vector badwords
filtf(x,myfilter) removes from character vector x regex matching items contained in vector myfilter
}

