% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/filts.R
\name{filts}
\alias{filts}
\title{filts.R}
\usage{
filts(x, f)
}
\arguments{
\item{x}{character}

\item{filters}{a char vector of tokens to exclude}
}
\description{
filts.R
}
\examples{
filts(as.character(citation),filters) removes from citation regex matching items contained in vector filters
filts(string,badwords) removes from string regex matching items contained in vector badwords
}

