% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/fr_currency_spell_words.R
\name{fr_currency_spell_words}
\alias{fr_currency_spell_words}
\title{fr_currency_spell_words}
\usage{
fr_currency_spell_words(x, fraction = FALSE, local = TRUE)
}
\arguments{
\item{x}{numeric}

\item{fraction}{logical if TRUE, convert also the fractional part}

\item{local}{logical if TRUE, compute locally, if FALSE use http service}
}
\value{
x as named num (name is the US_currency_spell_word_local)
}
\description{
fr_currency_spell_words
}
\examples{
fr_currency_spell_words(-101.45) returns the num -101.45 named "moins cent un euros"
fr_currency_spell_words(-101.45,TRUE) returns the num -101.45 named "moins cent un euros quarante-cinq euro-centimes"
fr_currency_spell_words(2001) returns the num 2001, named "deux mille un euros"
fr_currency_spell_words(2001.45,,TRUE) returns the num 2001.45, named "deux mille un euros"
fr_currency_spell_words(2001.45,TRUE,TRUE) returns the num 2001.45, named "deux mille un euros quarante-cinq euro-centimes"
fr_currency_spell_words(2001,,FALSE) returns the num 2001, named "The method is not implemented in the requested language."
fr_currency_spell_words(-128001000.78,TRUE,FALSE) returns the num -128001001, named "The method is not implemented in the requested language."
fr_currency_spell_words(-128001000.78,TRUE,TRUE) returns the num -128001001, named "moins cent vingt-huit million un mille euros six-dix-huit euro-centimes"
}

