% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/number_spell_words.R
\name{number_spell_words}
\alias{number_spell_words}
\title{number_spell_words}
\usage{
number_spell_words(n, l = "en_US")
}
\arguments{
\item{x}{integer number}
}
\value{
x as named integer (name is the number_spell_word)
}
\description{
number_spell_words
}
\examples{
number_spell_words(101) returns a named integer 101, name="one hundred one"
number_spell_words(-2001) returns a named integer -2001, name="minus two thousand one"
}

