% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/skip_ngram.R
\name{skip_ngram}
\alias{skip_ngram}
\title{skip_ngram}
\usage{
skip_ngram(x, n = 1L, skip = 0L, cflag = FALSE, filters = NULL)
}
\arguments{
\item{x}{character list, possibly compressed}

\item{n}{integer, the gram type (1L for unigram, 2L for bigram, etc..)}

\item{skip}{integer, the number of tokens to skip over (0L for simple gram, 1L for one skip, i.e [token X token]), note skip is ignored, set to 0L, on unigrams!}

\item{cflag}{logical, TRUE indicates compressed string x is input}

\item{filters}{is an optional character vector dictionary of tokens to exclude, omit if not supplied}
}
\value{
skip_ngram as named integer vector x
}
\description{
skip_ngram
}
\examples{
skip_ngram(as.character(citation())) produces a unigram of citation in list x
skip_ngram(string,3L) produces trigram of string in list x
skip_ngram(string,2L,3L,1L,badwords) produces skip3-bigram in list x corresponding to [word X X X word] from a compressed list, excludes tokens in badwords list
}

