% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/skip_ngram4.R
\name{skip_ngram4C}
\alias{skip_ngram4C}
\title{skip_ngram4C}
\usage{
skip_ngram4C(x, n = 1L, skip = 0L, cflag = 0L, filter = NULL)
}
\arguments{
\item{x}{character list, possibly compressed}

\item{n}{integer, the gram type}

\item{skip}{integer, the skip number}

\item{cflag}{integer, 1L if compressed string x is input}

\item{filter}{is an optional filter disctionary containing tokens to exclude}
}
\value{
n-gram in x
}
\description{
skip_ngram4C
}
\examples{
skip_ngram4C(as.character(citation())) produces a unigram of citation in list x
skip_ngram4C(string,3L) produces trigram of string in list x
skip_ngram4C(string,2L,3L,1L,badwords) produces skip3-bigram in list x corresponding to [word X X X word] from a compressed list, excludes tokens in badwords list
}

