% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/toCurrency.R
\name{toCurrency}
\alias{toCurrency}
\title{toCurrency Converts a currency to its word representation in a designated language locale}
\usage{
toCurrency(x, convert_fraction = TRUE, locale = "en_US")
}
\arguments{
\item{character}{string x, representing a decimal between -infinity and infinity to be converted in words}

\item{logical}{convert_fraction, indicates fraction to words (left as numeric if set to FALSE). Optional. Defaults to TRUE.}

\item{character}{string locale, representing the locale language, by default, set to 'us_EN'}
}
\value{
x as named num is the specified locale
}
\description{
toCurrency Converts a currency to its word representation in a designated language locale
}
\examples{
toCurrency('123.56') returns the num 123.56 named "one hundred twenty-three dollars fifty-six cents"
toCurrency('123.56',FALSE) returns the num 123.56 named "one hundred twenty-three dollars"
toCurrency('-2349.03',,'fr') returns the num -2349.03 named "moins deux mille trois cent quarante-neuf euros trois euro-centimes"
toCurrency('-2349.03',FALSE,'fr') returns the num -2349.03 named "moins deux mille trois cent quarante-neuf euros"
}

