% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/toCurrency_fr.R
\name{toCurrency_fr}
\alias{toCurrency_fr}
\title{toCurrency_fr Converts a currency value to its word representation (with monetary units) in French language}
\usage{
toCurrency_fr(decimal, fraction = FALSE, int_curr = "EUR",
  convert_fraction = TRUE)
}
\arguments{
\item{integer}{decimal, the currency total amount without fraction part (e.g. amount of dollars)}

\item{logical}{fraction, indicates if the fractional part of the currencyy amount (e.g. amount of cents) is desired. Optional. Defaults to FALSE.}

\item{integer}{int_curr, an international currency symbol as defined by the ISO 4217 standard (three characters)}

\item{logical}{convert_fraction indicates fraction to words conversion required (left as numeric if set to FALSE). Optional. Defaults to TRUE.}
}
\value{
string The corresponding word representation for the currency
}
\description{
toCurrency_fr Converts a currency value to its word representation (with monetary units) in French language
}
\examples{
toCurrency_fr('25','34') results in string "vingt-cinq euros trente-quatre euro-centimes"
toCurrency_fr('25','34',,FALSE) results in string "vingt-cinq euros 34 euro-centimes"
toCurrency_fr('-460','12','USD') results in string "moins quatre cent soixante dollars douze cents"
toCurrency_fr('+123','47','EUR') results in string "plus cent vingt-trois euros quarante-sept euro-centimes"
}

