% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/toWords.R
\name{toWords}
\alias{toWords}
\title{toWords Converts a number to its word representation in a designated language locale}
\usage{
toWords(x, convert_fraction = FALSE, locale = "en_US")
}
\arguments{
\item{character}{string x, representing a decimal between -infinity and infinity to be converted in words}

\item{logical}{convert_fraction, indicates fraction to words (left as numeric if set to FALSE). Optional. Defaults to TRUE.}

\item{character}{string locale, representing the locale language, by default, set to 'us_EN'}
}
\value{
x as named num is the specified locale
}
\description{
toWords Converts a number to its word representation in a designated language locale
}
\examples{
toWords('123.45') returns the num 123.45 named "one hundred twenty-three"
toWords('123.45',TRUE) returns the num 123.45 named "one hundred twenty-three and forty-five decimal parts"
toWords('123.45',,'fr') returns the num 123.45 named "cent vingt-trois"
toWords('-24123.45',,'fr') returns the num -24123.45 named "moins vingt-quatre mille cent vingt-trois"
toWords('-24123.45',TRUE,'fr') returns the num -24123.45 named "moins vingt-quatre mille cent vingt-trois et quarante-cinq parties decimales"
}

