% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/toWords_en_US.R
\name{toWords_en_US}
\alias{toWords_en_US}
\title{toWords_en_US Converts a number to its word representation in American English language}
\usage{
toWords_en_US(number, p = 0, powsuffix = "")
}
\arguments{
\item{character}{string number, representing an integer between -infinity and infinity to be converted in words}

\item{integer}{p, the power of ten for the rest of the number to the right. Used internally. Optional, defaults to 0.}

\item{character}{string powsuffix, the power name to be added to the end of the return string. Used internally. Optional, defaults to ''.}
}
\value{
string The corresponding word representation
}
\description{
toWords_en_US Converts a number to its word representation in American English language
}
\examples{
toWords_en_US('1234') returns the string "one thousand two hundred and thirty-four"
toWords_en_US('-678') returns the string "minus six hundred and seventy-eight"
toWords_en_US('-123456789012345678901234567890') returns the string "minus one hundred twenty-three octillion...hundred ninety"
toWords_en_US('+3','3','thousand') returns the string "plus three thousand"
}

