% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/toWords_fr.R
\name{toWords_fr}
\alias{toWords_fr}
\title{toWords_fr Converts a number to its word representation in French language}
\usage{
toWords_fr(number, p = 0, powsuffix = "")
}
\arguments{
\item{character}{string number, representing an integer between -infinity and infinity to be converted in words}

\item{integer}{p, the power of ten for the rest of the number to the right. Used internally. Optional, defaults to 0.}

\item{character}{string powsuffix, the power name to be added to the end of the return string. Used internally. Optional, defaults to ''.}
}
\value{
a character string containing the corresponding word representation
}
\description{
toWords_fr Converts a number to its word representation in French language
}
\examples{
toWords_fr('1234') returns the string "mille deux cent trente-quatre"
toWords_fr('-121') returns the string "moins trois cent vingt et un"
toWords_fr('+26') returns the string "plus vingt-six"
toWords_fr('+3','6','million') returns the string "plus trois mille"
toWords_fr('-1234567890123456789012345') returns the string "moins un septillion deux cent...trois cent quarante-cinq"
}

