% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/translate_number.R
\name{translate_number}
\alias{translate_number}
\title{translate_number}
\usage{
translate_number(x)
}
\arguments{
\item{x}{character}
}
\description{
translate_number
}
\examples{
translate_number( as.character( citation() ) ) substitutes number with spelled value
}

