% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/US_currency_spell_words.R
\name{US_currency_spell_words}
\alias{US_currency_spell_words}
\title{US_currency_spell_words}
\usage{
US_currency_spell_words(x, fraction = FALSE, local = TRUE)
}
\arguments{
\item{x}{numeric}

\item{fraction}{logical, by default TRUE, to convert also the fractional part}

\item{local}{logical, by default TRUE, to compute locally. if FALSE use http service}
}
\value{
x as named num (name is the US_currency_spell_word_local)
}
\description{
US_currency_spell_words
}
\examples{
US_currency_spell_words(101.45) returns the num 101.45 named "one hundred one dollars dollars and 45 cents"
US_currency_spell_words(-2001) returns the num -2001, named "minus two thousand one dollars"
US_currency_spell_words(2001.45,,TRUE) returns the num 2001.45, named "two thousand one dollars"
US_currency_spell_words(2001.45,TRUE,TRUE) returns the num 2001.45, named "two thousand one dollars forty-five cents"
US_currency_spell_words(2001.45,,FALSE) returns the num 2001.45, named "two thousand one dollars forty-five cents"
US_currency_spell_words(-128001000.78,TRUE,TRUE) returns the num -128001001, named "minus one hundred twenty-eight million one thousand dollars seventy-eight cents"
US_currency_spell_words(-128001000.78,TRUE,FALSE) returns the num -128001001, named "minus one hundred twenty-eight million one thousand dollars seventy-eight cents"
}

