% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/US_currency_spell_words.R
\name{US_currency_spell_words}
\alias{US_currency_spell_words}
\title{US_currency_spell_words}
\usage{
US_currency_spell_words(x)
}
\arguments{
\item{x}{numeric}
}
\value{
x as named num (name is the US_currency_spell_word)
}
\description{
US_currency_spell_words
}
\examples{
US_currency_spell_words(101.45) returns a named num 101.45 named "one hundred one dollars dollars and 45 cents"
US_currency_spell_words(-2001) returns a named num -2001, named "minus two thousand one dollars"
}

